requirements.txt
requests
jsonschema
pytest
allure-pytest
selenium
webdriver-manager